Rename benchmark-url to benchmark-server-timing.
