+++ cli/run.mjs
@@ -30,9 +30,9 @@ import {
 	formats,
 } from './lib/cli/logger.mjs';
 import {
-	handler as benchmarkUrlHandler,
-	options as benchmarkUrlOptions,
-} from './commands/benchmark-url.mjs';
+	handler as benchmarkServerTimingHandler,
+	options as benchmarkServerTimingOptions,
+} from './commands/benchmark-server-timing.mjs';
 import {
 	handler as wptMetricsHandler,
 	options as wptMetricsOptions,
@@ -66,9 +66,9 @@ const catchException = ( handler ) => {
 	};
 };
 
-withOptions( program.command( 'benchmark-url' ), benchmarkUrlOptions )
-	.description( 'Runs benchmarks for an URL' )
-	.action( catchException( benchmarkUrlHandler ) );
+withOptions( program.command( 'benchmark-server-timing' ), benchmarkServerTimingOptions )
+	.description( 'Runs Server Timing benchmarks for an URL' )
+	.action( catchException( benchmarkServerTimingHandler ) );
 withOptions( program.command( 'wpt-metrics' ), wptMetricsOptions )
 	.description( 'Gets performance metrics for a WebPageTest result' )
 	.action( catchException( wptMetricsHandler ) );
